Add files via upload
In this project we made a connection with XAMPP services for PHP and MySQL between our Database and our application. We did use Adapters (RecyclerView Adapters). Java classes for Averages, Courses, Login, Join in a course, Settings, SignUp, Userprofile, Comments, Posts, Validation of Users and Courses, Searching of courses.
JSON was a requeriment for SQL queries that extracts the information of the Database.